виправив помилку під час додавання нового контакту
