+++ models/contacts.json
@@ -67,13 +67,8 @@
   },
   {
     "id": "x98ntnlGryPNRiy9FSNSO",
-    "name": "viktor",
-    "email": "vvdiduh@fdjk.com",
-    "phone": 380979952939
-  },
-  {
-    "id": "lCSPmVnWUS7FfEioPLIwL",
-    "name": "vvvd3",
-    "email": "vvdiduh@fdjk.c1om"
+    "name": "sdfsdfsdf",
+    "email": "cc@dkgj.com",
+    "phone": "64654654"
   }
 ]
